package/libgpgme: add patch file to fix closefrom compile error
After a version bump of glibc, the following error occurs:

posix-io.c: In function '_gpgme_io_spawn':
posix-io.c:577:23: error: void value not ignored as it ought to be
  577 |             while ((i = closefrom (fd)) && errno == EINTR)
      |                       ^
make[3]: *** [Makefile:998: posix-io.lo] Error 1

It seems the closefrom function does not return any value anymore, and
is now a void type function.
In de code there is a check for platforms, and if the platform is __sun
or __FreeBSD__, the function is already called without assigning a
result from the function.

I simply added the platform on which it happened now (__linux) to this
check, which makes the code compile like before.